refactor: Seperate app config from server start
Each test file runs its own instance of the server. To make this possible I moved the server bootup step to be part of each test so you can specify a different port for each test runner
